requirements.txt
# Python dependencies for BMCP Realtime
# Core dependencies
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3

# Additional dependencies
pyyaml==6.0
pytest==7.3.1
pytest-cov==4.1.0

# Integration test dependencies
docker==6.1.3
neo4j==5.14.0
pika